🤝 make cloneObject work in both Node & browsers (#3571)
* Make cloneObject work in both Node & browsers

This fix ensures we check for `File` to be available as a global before we do an instance check. This is required as `File` isn't available on Node.js environments.

* Try to fix lint errors.

* Round #2 - Attempt to fix lint errors.

* Make cloneObject work in both Node & browsers

This fix ensures we check for `File` to be available as a global before we do an instance check. This is required as `File` isn't available on Node.js environments.

* clone-object: A better for the cloneObject() method.

Remove the isWeb parameter entirely and instead check it right inside the implementation.

* Update src/utils/cloneObject.ts

Include @bluebill1049's suggested fix.

Co-authored-by: Bill <bluebill1049@hotmail.com>

* cloneObject - Add missing import.

* move isWeb logic into its own file

Co-authored-by: Bill <bluebill1049@hotmail.com>